push new crud for categories

+++ database/seeders/DatabaseSeeder.php
@@ -3,6 +3,7 @@
 namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
+use Illuminate\Support\Facades\Schema;
 
 class DatabaseSeeder extends Seeder
 {
@@ -13,15 +14,23 @@ class DatabaseSeeder extends Seeder
     {
         $this->truncateTables();
         $this->call(AdminSeeder::class);
-        \App\Models\Shop::factory(5)->has(\App\Models\User::factory()->count(5))->create();
+        \App\Models\Shop::factory(5)
+                            ->has(\App\Models\User::factory()->count(5))
+                            ->has(\App\Models\Unit::factory()->count(5))
+                            ->has(\App\Models\Category::factory()->count(5))
+                            ->create();
     }
 
     protected function truncateTables()
     {
         $tableNames = \Illuminate\Support\Facades\Schema::getConnection()->getDoctrineSchemaManager()->listTableNames();
         foreach ($tableNames as $name) {
-            if ($name != 'migrations')
-                \Illuminate\Support\Facades\DB::table($name)->truncate();
+            if ($name != 'migrations' && $name != 'shops') {
+                Schema::enableForeignKeyConstraints();
+                    \Illuminate\Support\Facades\DB::table($name)->truncate();
+                Schema::disableForeignKeyConstraints();
+            }
         }
+        \Illuminate\Support\Facades\DB::table('shops')->truncate();
     }
 }
